+++ recipes/default.rb
@@ -85,6 +85,14 @@ when 'debian'
   end
 
 when 'rhel', 'fedora'
+  #This is needed since Erlang Solutions' packages provide "esl-erlang"; this package just requires "esl-erlang" and provides "erlang".
+  if node['erlang']['esl'] then
+    remote_file "#{Chef::Config[:file_cache_path]}/esl-erlang-compat.rpm" do
+      source "https://github.com/jasonmcintosh/esl-erlang-compat/blob/master/rpmbuild/RPMS/noarch/esl-erlang-compat-R14B-1.el6.noarch.rpm?raw=true"
+    end
+    rpm_package "#{Chef::Config[:file_cache_path]}/esl-erlang-compat.rpm"
+  end
+
   if node['rabbitmq']['use_distro_version'] then
     package 'rabbitmq-server'
   else